% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_evaluated_Network.R
\name{draw_evaluated_network}
\alias{draw_evaluated_network}
\title{Draw the result of a network evaluation}
\usage{
draw_evaluated_network(
  results,
  export = FALSE,
  filename = "./evaluated_network.png"
)
}
\arguments{
\item{results}{result returned by the function \code{evaluate_network} (of type list)}

\item{export}{weather or not to save the plot}

\item{filename}{name of the exported plot}
}
\value{
ggplot/ggraph object
}
\description{
Draw the result of a network evaluation
}
