% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateNetwork.R
\name{evaluate_network}
\alias{evaluate_network}
\title{Evaluates an inferred network against validated regulatory interactions}
\usage{
evaluate_network(
  net,
  validation = c("CHIPSeq", "DAPSeq", "TARGET"),
  subset_validated_edges = NULL
)
}
\arguments{
\item{net}{dataframe with a column 'from' (regulators) and a column 'to'
(targets), representing the inferred network of edges to evaluate}

\item{validation}{type of edge in the validation database that
should be considered to defined a true/supported prediction in the
evaluation process.
The validation type must be a vector of one or more
of the following values : CHIPSeq, DAPSeq, Litterature, TARGET}

\item{subset_validated_edges}{potential dataframe of validated edges
(that should however contain all the nodes in the inferred network).
Made to speed up computation in ROC and test again random.}
}
\value{
a list containing true positives, true positive rate,
false positives, false positive rate, and the input network dataframe
with an additional column to caracterise how is the edge supported
by known interactions
}
\description{
Evaluates an inferred network against validated regulatory interactions
}
