% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_edges.R
\name{flatten_edges}
\alias{flatten_edges}
\title{Deals with grouped nodes of a network}
\usage{
flatten_edges(net)
}
\arguments{
\item{net}{network containing grouped regulators,
in the form of a dataframe of edges. It must contain a column
named 'from' and a column names 'to'n with the AGI of the genes.
The grouped nodes must have the format mean_AGI1-AGI2, such as
returned by the inference in DIANE}
}
\value{
dataframe of edges with duplicated edges for
grouped regulators
}
\description{
The line :
from mean_ATGSJHKS-ATGODLKSL to ATGDHJKHK
}
\details{
becomes :

from ATGSJHKS to ATGDHJKHK
from ATGODLKSL to ATGDHJKHK
}
