% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_validation_rate.R
\name{test_validation_rate}
\alias{test_validation_rate}
\title{Test a network against random}
\usage{
test_validation_rate(
  net,
  validation = c("CHIPSeq", "DAPSeq", "TARGET"),
  N = 200
)
}
\arguments{
\item{net}{dataframe with a column 'from' (regulators) and a column 'to'
(targets), representing the inferred network of edges to evaluate}

\item{validation}{type of edge in the validation database that
should be considered to defined a true/supported prediction in the
evaluation process.
The validation type must be a vector of one or more
of the following values : CHIPSeq, DAPSeq, Litterature, TARGET}

\item{N}{Number of shuffled network to estimate the null ditribution
of the validation rate. 200 by default.}
}
\value{
A plot containing the empirical p-pvalue
of the permutation-based test
}
\description{
Test a network against random
}
\examples{
data("abiotic_stress_Heat_genes_net")
set.seed(999)
results <- evaluate_network(abiotic_stress_Heat_genes_net)
test_validation_rate(abiotic_stress_Heat_genes_net, N = 100)
}
