% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateNetwork.R
\name{evaluate_network}
\alias{evaluate_network}
\title{Evaluate Network}
\usage{
evaluate_network(
  net,
  validation = c("CHIPSeq", "DAPSeq", "Litterature", "TARGET"),
  subset_validated_edges = NULL
)
}
\arguments{
\item{net}{dataframe with a column 'from' (regulators) and a column 'to'
(targets), representing the inferred network of edges to evaluate}

\item{validation}{type of edge in the validation database that
should be considered to defined a true/supported prediction in the
evaluation process.
The validation type must be a vector of one or more
of the following values : CHIPSeq, DAPSeq, Litterature, TARGET}

\item{subset_validated_edges}{}
}
\value{
a list containing true positives, true positive rate,
false positives, false positive rate
}
\description{
Evaluates an inferred network against validated regulatory interactions
}
